ggsegExtra: Extra Utilities for the 'ggseg' and 'ggseg3d' Plotting Tools

Contains functions to create new data sets compatible with the 'ggseg' and 'ggseg3d' packages for plotting brain segmentations through R. Requires several external software to be installed on the system for atlas creations to work, and most of these are not available for windows. But atlas creation tools exist for 2d simple features data, and also 3d tri-surface mesh plots. Also has a list of compatible data sets and functions for easy installation of these.

See also

Author

Maintainer: Athanasia Mo Mowinckel a.m.mowinckel@psykologi.uio.no (ORCID)

Authors:

Didac Vidal-Piñeiro d.v.pineiro@psykologi.uio.no (ORCID)

Other contributors:

John Muschelli muschellij2@gmail.com (ORCID) [contributor]