Install all ggseg-atlases registeres

Calls ggseg_atlas_repos, and installs all libraries listed. Be careful calling this function, this will likely take quite some time to complete, and also will contain a lot of heavy data. We recommend only installing atlases you are likely to use and on demand.

install_ggseg_atlas_all(...)

Arguments

...
additional arguments to remotes::install_* function called depending on the source.

Examples

if (FALSE) {
install_ggseg_atlas_all()
}