Volumetric segmentation to 3d-atlas

Function to create a ggseg3d-atlas from a volumetric parcellation. Currently only tested using .mgz extension images. Will call command-line tools to complete the process.

make_volumetric_2_3datlas(
  subject = "fsaverage5",
  subjects_dir = freesurfer::fs_subj_dir(),
  template = file.path(subjects_dir, subject, "mri/aseg.mgz"),
  color_lut = file.path(freesurfer::fs_dir(), "ASegStatsLUT.txt"),
  steps = 1:5,
  output_dir = tempdir(),
  verbose = TRUE,
  ncores = parallel::detectCores() - 2,
  cleanup = TRUE
)

Arguments

subject
Freesurfer subject, must exist in whatever subject directory specified or set in the environment with $SUBJECTS_DIR

subjects_dir
Freesurfer subject directory

template
tempalte volume.mgz file path

color_lut
Freesurfer colour look-up-table. Either as a path or a data.frame that is_ctab

steps
if cleanup is disabled, all files are saved, and steps can be re-run individually

output_dir
output directory path

verbose
logical indicating to be verbose or not

ncores
number of cores for parallel processing (default numcores-2)

cleanup
logical to toggle removal of all intermediary files

Value

returns a ggseg3d-atlas ready object

Details

the function is aggressive in the number of cores used to minimize time taken to create the atlas. Set the ncores argument for less aggressive approach.